Remove context dependency from ProjectHeader
So we can re-use in V2
